abstract go specific code into languages/golang
